fix(code smells)Fix code climate smells   -Breakdown files that have more than 250 lines of code
