+++ fixtures/office/office_fixtures.py
@@ -129,93 +129,6 @@ office_mutation_query_duplicate_name_responce = {
     }
 }
 
-delete_office_mutation = '''
-mutation{
-    deleteOffice(officeId: 1){
-        office{
-            name
-            blocks{
-                name
-                floors{
-                    name
-                    rooms{
-                        name
-                        roomType
-                        capacity
-                        resources{
-                            name
-                            quantity
-                            devices{
-                                name
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-'''
-
-delete_office_mutation_response = {
-    "data": {
-        "deleteOffice": {
-            "office": {
-                "name": "St. Catherines",
-                "blocks": [
-                    {
-                        "name": "EC",
-                        "floors": [
-                            {
-                                "name": "3rd",
-                                "rooms": [
-                                    {
-                                        "name": "Entebbe",
-                                        "roomType": "meeting",
-                                        "capacity": 6,
-                                        "resources": [
-                                            {
-                                                "name": "Markers",
-                                                "quantity": 3,
-                                                "devices": [
-                                                    {
-                                                        "name": "Samsung"
-                                                    }
-                                                ]
-                                            }
-                                        ]
-                                    }
-                                ]
-                            }
-                        ]
-                    }
-                ]
-            }
-        }
-    }
-}
-
-delete_non_existent_office_mutation = '''
-
-mutation{
-    deleteOffice(officeId: 10){
-        office{
-            name
-        }
-    }
-}
-
-'''
-
-delete_unauthorised_location_mutation = '''
-mutation{
-    deleteOffice(officeId: 2){
-        office{
-            name
-        }
-    }
-}
-'''
 
 paginated_offices_query = '''
 query {
