+++ fixtures/room/filter_room_fixtures.py
@@ -1,5 +1,4 @@
-filter_rooms_by_capacity = '''
-query {
+filter_rooms_by_capacity = '''query {
   allRooms(capacity:6){
    rooms{
       name
@@ -24,8 +23,7 @@ filter_rooms_by_capacity_response = {
         }
     }
 }
-filter_rooms_by_location = '''
-query {
+filter_rooms_by_location = '''query {
   allRooms(location:"Kampala"){
    rooms{
       name
@@ -51,8 +49,7 @@ filter_rooms_by_location_response = {
     }
 }
 
-filter_rooms_by_resources = '''
-query {
+filter_rooms_by_resources = '''query {
   allRooms(resources:"Markers"){
    rooms{
       name
@@ -71,8 +68,7 @@ filter_rooms_by_resources_response = {
         }
     }
 }
-filter_rooms_by_resources_location_capacity = '''
-query {
+filter_rooms_by_resources_location_capacity = '''query {
   allRooms(location:"Kampala",capacity:6,resources:"Markers"){
    rooms{
       name
@@ -97,8 +93,7 @@ filter_rooms_by_resources_location_capacityresponse = {
         }
     }
 }
-filter_rooms_by_non_existant_data = '''
-query {
+filter_rooms_by_non_existant_data = '''query {
   allRooms(location:"kireka",capacity:7,resources:"Mghyt"){
    rooms{
       name
@@ -109,15 +104,14 @@ query {
     }
 }
     '''
-filter_rooms_by_non_existant_datay_response = {
+filter_rooms_by_non_existant_data_response = {
     "data": {
         "allRooms": {
             "rooms": []
         }
     }
 }
-filter_rooms_by_location_capacity = '''
-query {
+filter_rooms_by_location_capacity = '''query {
   allRooms(location:"Kampala",capacity:6){
    rooms{
       name
@@ -142,8 +136,7 @@ filter_rooms_by_location_capacity_response = {
         }
     }
 }
-filter_rooms_by_resources_capacity = '''
-query {
+filter_rooms_by_resources_capacity = '''query {
   allRooms(location:"Kampala",capacity:6,resources:"Markers"){
    rooms{
       name
@@ -168,8 +161,7 @@ filter_rooms_by_resources_capacity_response = {
         }
     }
 }
-filter_rooms_by_resources_location = '''
-query {
+filter_rooms_by_resources_location = '''query {
   allRooms(location:"Kampala",resources:"Markers"){
    rooms{
       name
@@ -195,8 +187,7 @@ filter_rooms_by_resources_location_response = {
     }
 }
 
-filter_rooms_by_office = '''
-query {
+filter_rooms_by_office = '''query {
   allRooms(office:"St. Catherines"){
    rooms{
       name
@@ -223,8 +214,7 @@ filter_rooms_response = {
     }
 }
 
-filter_rooms_by_office_capacity = '''
-query {
+filter_rooms_by_office_capacity = '''query {
   allRooms(office:"St. Catherines", capacity:6){
    rooms{
       name
@@ -235,8 +225,7 @@ query {
     }
 }
     '''
-filter_rooms_by_office_location = '''
-query {
+filter_rooms_by_office_location = '''query {
   allRooms(office:"St. Catherines", location:"Kampala"){
    rooms{
       name
@@ -247,8 +236,7 @@ query {
     }
 }
     '''
-filter_rooms_by_office_capacity_location = '''
-query {
+filter_rooms_by_office_capacity_location = '''query {
   allRooms(capacity:6 ,location:"Kampala", office:"St. Catherines"){
    rooms{
       name
