+++ tox.ini
@@ -1,7 +1,7 @@
 [tox]
 envlist = python2.7,
-    python 3.3,	
-    python3.4, 	
+    python 3.3, 
+    python3.4,  
     python3.5,	
     python3.6,	
     flake8
